+++ src/util/error.cpp
@@ -32,7 +32,8 @@ bilingual_str TransactionErrorString(const TransactionError error) {
             return Untranslated(
                 "Specified sighash value does not match existing value");
         case TransactionError::MAX_FEE_EXCEEDED:
-            return Untranslated("Fee exceeds maximum configured by -maxtxfee");
+            return Untranslated("Fee exceeds maximum configured by user (e.g. "
+                                "-maxtxfee, maxfeerate)");
             // no default case, so the compiler can warn about missing cases
     }
     assert(false);